Add support for partial matches when setting username.
Account can now be set with a case-insensitive comparison on username.

If an account is not found for the username as entered, an unambiguous match
will be selected, if one exists. Otherwise, the ambiguous matches will be
suggested in an error message.

This also adds a friendlier error if the username doesn't match any accounts.